refactor: share build alias loader
