+++ scripts/aliasUtils.mjs
@@ -1,0 +1,64 @@
+/**
+ * Utilities for deriving build-tool aliases from a package tsconfig.
+ */
+
+import { readFileSync } from 'fs';
+import { resolve } from 'path';
+
+export function loadAliases(rootDir) {
+  const tsconfig = readTsconfig(resolve(rootDir, 'tsconfig.json'));
+
+  return Object.fromEntries(
+    Object.entries(tsconfig.compilerOptions.paths)
+      .filter(([, values]) => !values[0].endsWith('.d.ts'))
+      .map(([key, values]) => {
+        const aliasKey = key.replace('/*', '');
+        const pathValue = values[0].replace('/*', '');
+
+        return [aliasKey, resolve(rootDir, pathValue)];
+      }),
+  );
+}
+
+function readTsconfig(tsconfigPath) {
+  const tsconfigText = readFileSync(tsconfigPath, 'utf8');
+  return JSON.parse(stripJsonComments(tsconfigText));
+}
+
+function stripJsonComments(text) {
+  let result = '';
+  let i = 0;
+  let inString = false;
+
+  while (i < text.length) {
+    const char = text[i];
+    const next = text[i + 1];
+
+    if (char === '"' && (i === 0 || text[i - 1] !== '\\')) {
+      inString = !inString;
+      result += char;
+      i++;
+      continue;
+    }
+
+    if (!inString) {
+      if (char === '/' && next === '/') {
+        while (i < text.length && text[i] !== '\n') i++;
+        continue;
+      }
+
+      if (char === '/' && next === '*') {
+        i += 2;
+        while (i < text.length - 1 && !(text[i] === '*' && text[i + 1] === '/'))
+          i++;
+        i += 2;
+        continue;
+      }
+    }
+
+    result += char;
+    i++;
+  }
+
+  return result.replace(/,(\s*[}\]])/g, '$1');
+}
